+++ depthmapXcli/agentparser.h
@@ -62,7 +62,7 @@ public:
 
 public:
     AgentParser();
-    virtual void parse(int argc, char *argv[]);
+    virtual void parse(size_t argc, char *argv[]);
     virtual void run(const CommandLineParser &clp, IPerformanceSink& perfWriter) const;
 
     enum AgentMode{
