+++ depthmapXcli/imodeparser.h
@@ -26,7 +26,7 @@ class IModeParser
 public:
     virtual std::string getModeName() const = 0;
     virtual std::string getHelp() const = 0;
-    virtual void parse( int argc, char **argv) = 0;
+    virtual void parse( size_t argc, char **argv) = 0;
     virtual void run(const CommandLineParser &clp, IPerformanceSink &perfWriter) const = 0;
     virtual ~IModeParser(){}
 };
