+++ depthmapXcli/segmentparser.cpp
@@ -46,7 +46,7 @@ std::string SegmentParser::getHelp() const {
            "Tulip)\n";
 }
 
-void SegmentParser::parse(int argc, char **argv) {
+void SegmentParser::parse(size_t argc, char **argv) {
     for (int i = 1; i < argc; ++i) {
         if (std::strcmp("-st", argv[i]) == 0) {
             if (m_analysisType != AnalysisType::NONE) {
