+++ depthmapXcli/vgaparser.cpp
@@ -25,7 +25,7 @@ using namespace depthmapX;
 VgaParser::VgaParser()
     : m_vgaMode(VgaMode::NONE), m_localMeasures(false), m_globalMeasures(false) {}
 
-void VgaParser::parse(int argc, char *argv[]) {
+void VgaParser::parse(size_t argc, char *argv[]) {
     for (int i = 1; i < argc;) {
 
         if (std::strcmp("-vm", argv[i]) == 0) {
