+++ depthmapXcli/visprepparser.h
@@ -44,7 +44,7 @@ public:
                "  -pl Remove links when unmaking\n";
     }
 
-    virtual void parse(int argc, char** argv);
+    virtual void parse(size_t argc, char** argv);
 
     virtual void run(const CommandLineParser &clp, IPerformanceSink &perfWriter) const;
