+++ modules/segmentshortestpaths/cli/segmentshortestpathparser.h
@@ -34,7 +34,7 @@ class SegmentShortestPathParser : public IModeParser {
 
     enum class StepType { NONE, TULIP, METRIC, TOPOLOGICAL };
 
-    virtual void parse(int argc, char **argv);
+    virtual void parse(size_t argc, char **argv);
 
     virtual void run(const CommandLineParser &clp, IPerformanceSink &perfWriter) const;
